refactor(character): extract JSON export serialization into utils file
The export button's script exceeded the 30-line threshold; the
serialization, exportSource stamping, and filename assembly move to a
co-located CharacterJsonExportButton.utils.ts (single caller, per the
code promotion rules), leaving the component as a thin UI shell that
handles only the click and user notifications.
